changed the way that the destination display works, so that, if section a box name is already selected as an option, if the section selection is changed, then the box options will reset to 'select' and the destination output for box will disappear, until user chooses a new box that belongs to the newly selected section.
Prior to this, the output would not change and this potentially causes confusing, where you would have the select menus and the output display showing different values for section and box names.

I've also made the change so that if both section and box are selected (when an item is being transferred), if a new location is selected, both section and box options will reset, and those parts of the destination display will disappear, waiting to be re-rendered once user selects those again.

If a box is being transferred, only the section select will be avialable and when user changes location after having selected a section, the section menu will reset and the previous section selection will disappear from the destination display, waiting for user to choose a section that corresponds to the newly selected location; since the sections menu automatically changes to match the new location that has been clicked.