currently trying to resolve issues with the transferring a box to a new section that already has a box with the same name as the intended transfer box.
I've gotten to the point where I can transfer the box after renaming it on after the 'duplicate' warning appears and an input field appears allowing the name change; the only part left which I'm working on is changing the color of the transferred box so that it displays in a green colour as a visual confirmation that of the actual box that was transferred.

This part of the process works properly where no duplicate box exists at the destination; the transferred box name is highlighted green.  But when there is a duplicate involved, even after renaming, when the transfer completes, the wrong box name is highlighted.  The renamed box which is the transfer box does not get highlighted, but the box which was already in the destination with the duplicate name is highlighted.

The function is highlighting the old name, the name that the transfer box had prior to renaming; but because that name existed in the destination also, since we are looping through the destination box to find the name that matches the original, it's the box that has the original name (i.e. the duplicate box at the destination) that gets its name highlighted.

what needs to happens that the map needs to compare the destination box names to the 'modified' name.  I've stored this in local storage so this might fix the issue.  I'm using local storage because I don't yet know how to do it with REACT alone; I was saving it in state but I think the page is re-rendering so the 'modified' name variable saved in state returns to an empty string, which is its original state prior to modification, so it is ignored and the old name value is used.