after duplicate rename, transfer box with new name is highlighted as is supposed to happen.
Also, I found the reason why the modified name in state was no longer available after applying the transfer. It was reset in the function which actually moves the box to its new destination.  I'll leave it there for now. But I will likely revert to using it as long as it is available after transfer.

Local storage will need to be cleared at some point after the transfer.  There are three cases where this should happen.  The only ways to escape the transfer page post transfer are, a) via the 'go to origin' button which takes you to the source of the transfer box, b) via the 'go to destination' button which takes you to the destination of the box, or, c) refresh the applications.  So I will clear the local storage 'modified name' variable in the functions executed by the origin and destination buttons, and on the main app page so that if the webpage is refreshed, that will clear the modified name variable.

but until that happens I will use the variable to display the rename in the heading that shows which box was transferred.